+++ gzhttp/writer/gzstd/stdlib.go
@@ -1,4 +1,4 @@
-package gzkp
+package gzstd
 
 import (
 	"compress/gzip"
