+++ gzhttp/writer/gzstd/stdlib_test.go
@@ -1,4 +1,4 @@
-package gzkp
+package gzstd
 
 import (
 	"bytes"
